+++ pkg/domain/vulnerability.go
@@ -164,76 +164,3 @@ type CWEDetail struct {
 	CreatedAt   *time.Time `json:"created_at"`
 	LastUpdated *time.Time `json:"last_updated"`
 }
-
-// GeneratePortDataFromWebVuln creates PortData based on WebVulnerability information
-func GeneratePortDataFromWebVuln(vuln tools.WebVulnerability) tools.PortData {
-	portData := tools.PortData{
-		ID:       80,
-		Protocol: "tcp",
-		Service: tools.Service{
-			Name: "http",
-		},
-		State:           "open",
-		Vulnerabilities: nil,
-	}
-	if vuln.WascID == "45" {
-		portData.ID = 443
-		portData.Protocol = "tcp"
-		portData.Service.Name = "nginx"
-	}
-	return portData
-}
-
-// GenerateVulnFromWebVuln creates a tools.Vulnerability from WebVulnerability and Scan data
-func GenerateVulnFromWebVuln(scan *Scan, vuln tools.WebVulnerability) tools.Vulnerability {
-	today := time.Now()
-	return tools.Vulnerability{
-		HostID:             scan.HostID,
-		ScanID:             scan.ID,
-		CweID:              vuln.CweID,
-		Type:               enums.OwaspCategoryNoInfo,
-		BaseCVSSScore:      0,
-		References:         []string{vuln.Reference},
-		Description:        vuln.Name,
-		Access:             enums.AccessTypeUnknown,
-		Complexity:         enums.ComplexityTypeUnknown,
-		PrivilegesRequired: enums.PrivilegesRequiredUnknown,
-		Likelihood:         enums.LikelyhoodTypeUnknown,
-		RiskScore:          0,
-		ImpactScore:        0,
-		Exploit:            tools.Exploit{},
-		IntegrityImpact:    enums.ImpactTypeUnknown,
-		AvailabilityImpact: enums.ImpactTypeUnknown,
-		BaseSeverity:       enums.SeverityTypeUnknown,
-		Published:          today,
-		LastUpdated:        today,
-	}
-}
-
-func GenerateCWEDetailFromVuln(vuln tools.Vulnerability) CWEDetail {
-	// Extract title from first CWERemediation record
-	title := ExtractCWETitleFromRemediation(vuln.CWERemediation)
-	now := time.Now()
-	return CWEDetail{
-		ID:          vuln.CweID,
-		Title:       title,
-		CreatedAt:   &now,
-		LastUpdated: &now,
-	}
-}
-
-func GenerateCWEDetailFromWebVuln(vuln tools.WebVulnerability) CWEDetail {
-	now := time.Now()
-	return CWEDetail{
-		ID:          vuln.CweID,
-		CreatedAt:   &now,
-		LastUpdated: &now,
-	}
-}
-
-func ExtractCWETitleFromRemediation(remediations []tools.CWERemediation) string {
-	if len(remediations) > 0 {
-		return remediations[0].Title
-	}
-	return ""
-}
