+++ pkg/services/vulnerability.go
@@ -5,6 +5,7 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
+	"github.com/kptm-tools/core-service/pkg/services/helpers"
 	"log/slog"
 	"time"
 
@@ -80,7 +81,7 @@ func (s *VulnerabilityService) CreateNetworkOSVulnerabilities(
 				return fmt.Errorf("failed to insert os vulnerability CVE detail: %w", errCreateCVE)
 			}
 			// 2.1.2 Store CWE detail of the vuln
-			_, errCreateCWE := s.cweRepo.CreateOrUpdateCWE(ctx, domain.GenerateCWEDetailFromVuln(vuln))
+			_, errCreateCWE := s.cweRepo.CreateOrUpdateCWE(ctx, helpers.GenerateCWEDetailFromVuln(vuln))
 			if errCreateCWE != nil {
 				return fmt.Errorf("failed to insert os vulnerability CWE detail: %w", errCreateCWE)
 			}
@@ -118,7 +119,7 @@ func (s *VulnerabilityService) CreateNetworkOSVulnerabilities(
 				if errCreateCVE != nil {
 					return fmt.Errorf("failed to insert port vulnerability CVE detail: %w", errCreateCVE)
 				}
-				_, errCreateCWE := s.cweRepo.CreateOrUpdateCWE(ctx, domain.GenerateCWEDetailFromVuln(vuln))
+				_, errCreateCWE := s.cweRepo.CreateOrUpdateCWE(ctx, helpers.GenerateCWEDetailFromVuln(vuln))
 				if errCreateCWE != nil {
 					return fmt.Errorf("failed to insert os vulnerability CWE detail: %w", errCreateCWE)
 				}
@@ -375,7 +376,7 @@ func (s *VulnerabilityService) InsertWebScanVulnerabilities(ctx context.Context,
 		// 2.1 Loop through each webScan vulnerability, saving each one
 		for _, vuln := range webScanResult.WebVulnerabilities {
 			// 2.1.0 Store the default service
-			portData := domain.GeneratePortDataFromWebVuln(vuln)
+			portData := helpers.GeneratePortDataFromWebVuln(vuln)
 			// check if exists service
 			service, errGetExistence := s.serviceRepo.GetServiceByScanHostPortProtocol(ctx, scan.HostID, scan.ID, int32(portData.ID), portData.Protocol)
 			if errGetExistence != nil && !errors.Is(errGetExistence, customerrors.ErrServiceNotFound) {
@@ -388,7 +389,7 @@ func (s *VulnerabilityService) InsertWebScanVulnerabilities(ctx context.Context,
 				}
 			}
 			// 2.1.1 Store CWE detail of the vuln
-			_, err = s.cweRepo.CreateOrUpdateCWE(ctx, domain.GenerateCWEDetailFromWebVuln(vuln))
+			_, err = s.cweRepo.CreateOrUpdateCWE(ctx, helpers.GenerateCWEDetailFromWebVuln(vuln))
 			if err != nil {
 				return fmt.Errorf("failed to insert web scan vulnerability CWE detail: %w", err)
 			}
@@ -399,7 +400,7 @@ func (s *VulnerabilityService) InsertWebScanVulnerabilities(ctx context.Context,
 				return fmt.Errorf("failed to insert web scan vulnerability WASC detail: %w", err)
 			}
 
-			vulnData := domain.GenerateVulnFromWebVuln(scan, vuln)
+			vulnData := helpers.GenerateVulnFromWebVuln(scan, vuln)
 			// 2.1.3 Create a Vulnerability Record
 			dbVuln, err := s.vulnerRepo.CreateVulnerability(ctx, scan.HostID, scan.ID, vulnData, repository.VulnerabilityTypeEnumWEBAPPLICATION, "OWASP ZAP", "")
 			if err != nil {
